requirements.txt
torchnet==0.0.4
tqdm==4.40.1